\name{PP34b}
\alias{PP34b}
\title{A P-I model dataset}
\description{A model dataset that ...
  \itemize{
    \item blaba,
    \item blabal
    \item etc...
  }

}
\usage{PP34b}
\author{Tom Cox <t.cox@nioo.knaw.nl>}
\examples{
## ======================
## the model parameters:
## ======================
  
}
\references{

????
}
\details{

}
\seealso{
  Other data sets not yet included
}
\keyword{data}

