\name{primprod-package}
\alias{primprod-package}
\alias{primprod}
\docType{package}
\title{
  Photosynthesis - irradiance modelling
}
\description{
  Assessing phytoplankton community photosynthesi seasonality and response
  to dissolved silica depletion from photosynthesis-irradiance curves
}
\details{
  \tabular{ll}{
    Package: \tab primprod\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    License: \tab GNU Public License 2 or above\cr
  }

  primprod contains functions to fit P-E curves, adding co-limiting
  factors (temperature, Si concentration) and provides functions to assess
  the best-fit model
}
\author{
  Tom Cox (maintainer)

  Karline Soetaert
}
%\references{
%% to be filled
%}
\seealso{
 \code{\link{Onestep}}, the one-step functions

 \code{\link{Twostep}}, the two-step functions

}

\examples{
\dontrun{
## show examples (see respective help pages for details)
example(Twostep)
example(Onestep)
}
}

\keyword{ package }